+++ yasio/bindings/yasio_jsb20.hpp
@@ -1,5 +1,5 @@
 //////////////////////////////////////////////////////////////////////////////////////////
-// A multi-platform support c++11 library with focus on asynchronous socket I/O for any 
+// A multi-platform support c++11 library with focus on asynchronous socket I/O for any
 // client application.
 //////////////////////////////////////////////////////////////////////////////////////////
 /*
@@ -45,4 +45,3 @@ void clear();
 } // namespace yasio_jsb
 
 #endif
-